requirements.txt
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
boto3==1.26.83
botocore==1.29.83
certifi==2022.12.7
charset-normalizer==3.1.0
#decorator==5.1.1
executing==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
ipython==8.12.0
jedi==0.18.2
jmespath==1.0.1
matplotlib-inline==0.1.6
moviepy==1.0.3
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-autoremove==0.10.0
proglog==0.1.10
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
s3transfer==0.6.0
sentry-sdk==1.15.0
shapely==2.0.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
wcwidth==0.2.6